10-flask-revision/pyproject.toml
[project]
name = "10-flask-revision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "mysql-connector-python>=9.3.0",
]
